requirements.txt
aniso8601==1.1.0
click==6.6
decorator==4.0.10
Flask==0.11.1
Flask-RESTful==0.3.5
Flask-Restless==0.17.0
Flask-SQLAlchemy==2.1
ipython==5.1.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mimerender==0.6.0
pexpect==4.2.0
pickleshare==0.7.4
prompt-toolkit==1.0.6
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==2.5.3
python-mimeparse==1.5.2
pytz==2016.6.1
requests==2.11.1
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.0.14
traitlets==4.2.2
wcwidth==0.1.7
Werkzeug==0.11.10
